requirements.txt
# Core dependencies
pyaudio==0.2.13
pydub==0.25.1
numpy==1.24.3
protobuf==4.24.3
grpcio==1.58.0
python-dotenv==1.0.0

# LangChain ecosystem instead of direct OpenAI
langchain==0.1.20
langchain-openai==0.1.8
langchain-core==0.1.52
langchain-community==0.0.38

# For future RAG and vector stores
langchain-chroma==0.1.2
chromadb==0.4.24

# Audio processing
ffmpeg-python==0.2.0

# Project structure
setuptools==68.2.2
wheel==0.41.2

# Testing
pytest==7.4.2
pytest-mock==3.11.1